Remove the default argument for 'filter'
The default argument is somewhat confusing. This function is only called in one place, line 176. In that place, we build the list filter, using `position = 'none', clear = TRUE, plain = FALSE` as defaults. However, here they can be individually overridden. To call this function with no argument for filter would not make sense, as it would return without doing anything.